requirements.txt
django==3.2
djangorestframework==3.4.1
pyjwt==2.3.0
